backend/requirements.txt
fastapi>=0.111.1
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
Pillow>=10.4.0
fal-client>=0.13.0
httpx>=0.27.0
